+++ packages/code-components/src/test/setup.ts
@@ -1,0 +1,11 @@
+import { afterEach } from 'vitest';
+
+declare global {
+  var IS_REACT_ACT_ENVIRONMENT: boolean;
+}
+
+globalThis.IS_REACT_ACT_ENVIRONMENT = true;
+
+afterEach(() => {
+  document.body.innerHTML = '';
+});